Add AppStream .metainfo.xml file for Linux app stores
I had to make a number of assumptions here, like what the ID and
metainfo.xml license should be, whether to include the releases, etc. I
think it will be easier to talk about them in the pull request, though,
as then I'll be able to highlight the relevant lines for each point.